[ember] @ember/routing as source of types, ember as re-export
