+++ types/ember/tsconfig.json
@@ -24,6 +24,8 @@
             "@ember/array/*": ["ember__array/*"],
             "@ember/debug": ["ember__debug"],
             "@ember/debug/*": ["ember__debug/*"],
+            "@ember/routing": ["ember__routing"],
+            "@ember/routing/*": ["ember__routing/*"],
             "@ember/object": ["ember__object"],
             "@ember/object/*": ["ember__object/*"],
             "@ember/component": ["ember__component"],
