+++ types/ember__routing/hash-location.d.ts
@@ -1,3 +1,8 @@
-import Ember from 'ember';
+import EmberObject from "@ember/object";
 
-export default class HashLocation extends Ember.HashLocation { }
+/**
+ * `Ember.HashLocation` implements the location API using the browser's
+ * hash. At present, it relies on a `hashchange` event existing in the
+ * browser.
+ */
+export default class HashLocation extends EmberObject {}
