+++ types/ember__routing/index.d.ts
@@ -6,3 +6,9 @@
 
 export { default as Route } from '@ember/routing/route';
 export { default as Router } from '@ember/routing/router';
+import RouterService from '@ember/routing/router-service';
+
+// tslint:disable-next-line:strict-export-declare-modifiers
+interface Registry {
+    'router': RouterService;
+}
