+++ types/ember__routing/location.d.ts
@@ -1,4 +1,13 @@
-import Ember from 'ember';
-
-export const Location: typeof Ember.Location;
+/**
+ * Ember.Location returns an instance of the correct implementation of
+ * the `location` API.
+ */
+declare const Location: {
+    /**
+     * This is deprecated in favor of using the container to lookup the location
+     * implementation as desired.
+     * @deprecated Use the container to lookup the location implementation that you need.
+     */
+    create(options?: {}): any;
+};
 export default Location;
