+++ types/ember__routing/test/router.ts
@@ -1,7 +1,9 @@
-import Ember from 'ember';
 import { assertType } from './lib/assert';
+import Router from '@ember/routing/router';
+import Service, { inject as service } from '@ember/service';
+import EmberObject, { get } from '@ember/object';
 
-const AppRouter = Ember.Router.extend({
+const AppRouter = Router.extend({
 });
 
 AppRouter.map(function() {
@@ -24,31 +26,31 @@ AppRouter.map(function() {
     this.mount('my-engine', { as: 'some-other-engine', path: '/some-other-engine'});
 });
 
-const RouterServiceConsumer = Ember.Service.extend({
-    router: Ember.inject.service('router'),
+const RouterServiceConsumer = Service.extend({
+    router: service('router'),
     currentRouteName() {
-        const x: string = Ember.get(this, 'router').currentRouteName;
+        const x: string = get(this, 'router').currentRouteName;
     },
     currentURL() {
-        const x: string = Ember.get(this, 'router').currentURL;
+        const x: string = get(this, 'router').currentURL;
     },
     transitionWithoutModel() {
-        Ember.get(this, 'router')
+        get(this, 'router')
         .transitionTo('some-route');
     },
     transitionWithModel() {
-        const model = Ember.Object.create();
-        Ember.get(this, 'router')
+        const model = EmberObject.create();
+        get(this, 'router')
         .transitionTo('some.other.route', model);
     },
     transitionWithMultiModel() {
-        const model = Ember.Object.create();
-        Ember.get(this, 'router')
+        const model = EmberObject.create();
+        get(this, 'router')
         .transitionTo('some.other.route', model, model);
     },
     transitionWithModelAndOptions() {
-        const model = Ember.Object.create();
-        Ember.get(this, 'router')
+        const model = EmberObject.create();
+        get(this, 'router')
         .transitionTo('index', model, { queryParams: { search: 'ember' }});
     }
 });
